rename counter route to receive
