+++ app/Routes.js
@@ -3,7 +3,7 @@ import { Switch, Route } from 'react-router';
 import routes from './constants/routes';
 import App from './containers/App';
 import HomePage from './containers/HomePage';
-import CounterPage from './containers/CounterPage';
+import ReceivePage from './containers/ReceivePage';
 import SendPage from './containers/SendPage';
 import ImportPage from './containers/ImportPage';
 import ImportKeyPage from './containers/ImportKeyPage';
@@ -15,7 +15,7 @@ import FirstStartupPage from './containers/FirstStartupPage';
 export default () => (
   <App>
     <Switch>
-      <Route exact path={routes.COUNTER} component={CounterPage} />
+      <Route exact path={routes.RECEIVE} component={ReceivePage} />
       <Route exact path={routes.HOME} component={HomePage} />
       <Route exact path={routes.SEND} component={SendPage} />
       <Route exact path={routes.IMPORT} component={ImportPage} />
