+++ app/reducers/index.js
@@ -1,7 +1,7 @@
 // @flow
 import { combineReducers } from 'redux';
 import { connectRouter } from 'connected-react-router';
-import counter from './counter';
+import counter from './receive';
 
 export default function createRootReducer(history: History) {
   return combineReducers<{}, *>({
