+++ app/store/configureStore.dev.js
@@ -4,7 +4,7 @@ import { createHashHistory } from 'history';
 import { routerMiddleware, routerActions } from 'connected-react-router';
 import { createLogger } from 'redux-logger';
 import createRootReducer from '../reducers';
-import * as counterActions from '../actions/counter';
+import * as counterActions from '../actions/receive';
 import type { counterStateType } from '../reducers/types';
 
 const history = createHashHistory();
